==> 02. Cleaning and n-gram generation (clumsy code).ipynb <==
import numpy as np
import sklearn
import matplotlib.pyplot as plt
%pylab inline
pylab.rcParams['figure.figsize'] = (20, 10)
plt.style.use('fivethirtyeight')
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer

### Data cleaning

#### Tweets

with open('en_US/en_US.twitter.txt') as myfile:
    tweets = [next(myfile) for x in myfile]

tweets[0:5]

string = ' '.join(tweets)

# remove ascii
string = unicode(string, 'ascii', 'ignore')

string[500:1000]

# remove numbers
import re
string = re.sub(r'\d+', '', string)

string[500:1000]

# remove special characters
#string = re.sub('[^A-Za-z0-9]+', ' ', string)

string = string.split('\n')

string[0:10]

# cleaning special characters
#cleaned = []
#for i in arange(len(string)):
#    cleaned.append(re.sub("(@[A-Za-z0-9]+)|([^0-9A-Za-z \t])|(\w+:\/\/\S+)", ' ', string[i]))

cleaned[0:10]

len(cleaned)*0.3

### n-gram Generation

from nltk.tokenize import TreebankWordTokenizer

def ngram_freq_generator(string,n,features_no):
    cv = CountVectorizer(ngram_range=(n,n),max_features=features_no)
    s = cv.fit_transform(string)
    X = s.todense()
    features = cv.get_feature_names()
    counts = np.asarray(X.sum(axis=0)).ravel().tolist()
    grams_df = pd.DataFrame({'term': features, 'occurrences': np.asarray(counts)})
    grams_df = grams_df.sort_values('occurrences',ascending=False)
    return grams_df

# only using 30% of lines
bigrams = ngram_freq_generator(string[0:354022],2,5000)

bigrams.to_csv('bigrams_twitter.csv', index= False)

onegrams = ngram_freq_generator(string[0:354022],1,5000)

onegrams.to_csv('onegrams_twitter.csv', index= False)

trigrams = ngram_freq_generator(string[0:354022],3,5000)

trigrams.to_csv('trigrams_twitter.csv', index= False)

### Data cleaning

#### Blogs

with open('en_US/en_US.blogs.txt') as myfile:
    blogs = [next(myfile) for x in myfile]

len(blogs)

blogs[0:5]

string = ' '.join(blogs)
string = unicode(string, 'ascii', 'ignore')
string = re.sub(r'\d+', '', string)
string = string.split('\n')

len(string)*0.5

bigrams = ngram_freq_generator(string[0:224823],2,5000)

bigrams.to_csv('bigrams_blogs.csv', index= False)

trigrams = ngram_freq_generator(string[0:224823],3,5000)

trigrams.to_csv('trigrams_blogs.csv', index= False)

onegrams = ngram_freq_generator(string[0:224823],3,5000)

onegrams.to_csv('onegrams_blogs.csv', index= False)

import re

def string_clean(string):
    string = ' '.join(string)
    string = unicode(string, 'ascii', 'ignore')
    string = re.sub(r'\d+', '', string)
    string = string.split('\n')
    return string    

def write_grams(string,gram,names):
    onegrams = ngram_freq_generator(string,gram[0],5000)
    onegrams.to_csv(names[0],index=False)
    bigrams = ngram_freq_generator(string,gram[1],5000)
    bigrams.to_csv(names[1],index=False)
    trigrams = ngram_freq_generator(string,gram[2],5000)
    trigrams.to_csv(names[2],index=False)

with open('en_US/en_US.news.txt') as myfile:
    news = [next(myfile) for x in xrange(77259)]

news[0:5]

cleaned = string_clean(news)

cleaned[0:5]

gram = [1,2,3]
names = ['onegrams_news.csv','bigrams_news.csv','trigrams_news.csv']

write_grams(cleaned,gram,names)
