==> 04. Features generation from individual ngrams.ipynb <==
import numpy as np
import sklearn
import matplotlib.pyplot as plt
%pylab inline
pylab.rcParams['figure.figsize'] = (20, 10)
plt.style.use('fivethirtyeight')
import pandas as pd
import re

# read all bigrams
b_blogs = pd.read_csv('bigrams_blogs.csv')
b_news = pd.read_csv('bigrams_news.csv')
b_twit = pd.read_csv('bigrams_twitter.csv')

bigrams = b_blogs.append(b_news,ignore_index= True).append(b_twit,ignore_index= True)

We now have 15000 bigrams with their frequencies. It would be safe to assume that some of the bigrams generated from each source will be same. We will use pandas groupby method to aggregate the data. Precisely speaking, we will add the occurrences if the terms match.

bigrams.head()

g = bigrams.groupby('term').sum()

g.head()

len(g)

As we can see, 15000 features got aggregated in to 8800

g['term'] = g.index

g.head()

g = g.reset_index(drop=True)

g.head()

g = g.sort_values('occurrences',ascending=False)

g.head()

g.tail()

g.to_csv('bigrams_aggregated.csv', index=False)

def gen_agg(g,name):
    g = g.groupby('term').sum()
    g['term'] = g.index
    g = g.reset_index(drop=True)
    g = g.sort_values('occurrences',ascending=False)
    g.to_csv(name, index=False)

t_blogs = pd.read_csv('trigrams_blogs.csv')
t_news = pd.read_csv('trigrams_news.csv')
t_twit = pd.read_csv('trigrams_twitter.csv')

trigrams = t_blogs.append(t_news,ignore_index= True).append(t_twit,ignore_index= True)

gen_agg(trigrams,'trigrams_aggregated.csv')

q_blogs = pd.read_csv('fourgrams_blogs.csv')
q_news = pd.read_csv('fourgrams_news.csv')
q_twit = pd.read_csv('fourgrams_tweets.csv')

fourgrams = q_blogs.append(q_news,ignore_index= True).append(q_twit,ignore_index= True)

gen_agg(fourgrams,'fourgrams_aggregated.csv')

o_blogs = pd.read_csv('onegrams_blogs.csv')
o_news = pd.read_csv('onegrams_news.csv')
o_twit = pd.read_csv('onegrams_twitter.csv')

onegrams = o_blogs.append(o_news,ignore_index= True).append(o_twit,ignore_index= True)

gen_agg(onegrams,'onegrams_aggregated.csv')
